+++ forge/src/main/java/com/jahirtrap/configlib/ConfigLibMod.java
@@ -7,14 +7,21 @@ import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
-@Mod("configlibtxf")
-@Mod.EventBusSubscriber(modid = "configlibtxf", bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+@Mod(ConfigLibMod.MODID)
+@Mod.EventBusSubscriber(modid = ConfigLibMod.MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class ConfigLibMod {
+
+    public static final String MODID = "configlibtxf";
+
+    public ConfigLibMod() {
+        TXFConfig.init(MODID, ExampleConfig.class, "example");
+    }
+
     @SubscribeEvent
     public static void onClientSetup(FMLClientSetupEvent event) {
         System.setProperty("java.awt.headless", "false");
         ModList.get().forEachModContainer((modid, container) -> {
-            if (TXFConfig.configClass.containsKey(modid)) {
+            if (TXFConfig.configClass.keySet().stream().anyMatch(k -> k.equals(modid) || k.startsWith(modid + ":"))) {
                 container.registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () ->
                         new ConfigScreenHandler.ConfigScreenFactory((client, parent) -> TXFConfigClient.getScreen(parent, modid)));
             }